Controllers 1. Added controllers folder and respective controlthe logic 2. Changes to routes to incorporate the controllers
